feat: create product component

+++ shared/Models/Product.cs
@@ -12,8 +12,8 @@ namespace shared.Models
         public string Name { get; set; }
         public string Description { get; set; }
         public string Image { get; set; }
-        public float Price { get; set; }
-        public float? Discount { get; set; }
+        public decimal Price { get; set; }
+        public decimal? Discount { get; set; }
         public Collection<OrderItem> OrderItem { get; set; }
     }
 }
